+++ cedar-policy/CHANGELOG.md
@@ -29,6 +29,7 @@ Cedar Language Version: 4.0
   typename that can resolve to either an entity or common type, matching the
   behavior of typenames written in the human-readable (Cedar) syntax. (#1060, as
   part of resolving #579)
+- Add convenience methods to see how many policies and templates a policy set has (#1179)
 
 ### Changed